+++ forttask/src/components/generalUI/calendar.tsx
@@ -31,7 +31,7 @@ export default function Calendar({initialDate = new Date(), onChange}: CalendarP
     const endDate = endOfWeek(monthEnd);
     const days = eachDayOfInterval({ start: startDate, end: endDate });
 
-    const styleForSelected = "bg-[#A1A1AA] text-[#18181B] font-semibold hover:bg-[#A1A1AA] hover:text-[#FAFAFA]";;
+    const styleForSelected = "bg-[#A1A1AA] text-[#18181B] font-semibold hover:bg-[#A1A1AA] hover:text-[#FAFAFA]";
 
     const handleDateClick = (day: Date) => {
         if (isSameMonth(day, monthStart) && (isToday(day) || !isBefore(day, today))) {
